pyproject.toml
[tool.poetry]
name = "demo-qa"
version = "0.1.0"
description = ""
authors = ["Maksim Kudaev <kudaev.m@nakukop.com>"]
readme = "README.md"
packages = [{include = "demo_qa"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.2"
allure-pytest = "^2.13.0"
selene = "^2.0.0b17"
dynaconf = "^3.1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = " -vv -s --alluredir=allure-results "
